Create reference lines plotly as the first stage of interactive plot

Create reference lines plotly as the first stage of interactive plot

create_reference_lines_for_plotly(
  performance_table_type,
  curve,
  prevalence = NA,
  population_color_vector = NA
)

Arguments

performance_table_type
the type of the performance table

curve
the required curve

prevalence
the prevalence of the population

population_color_vector
color values

Value