Set Titles for x and y axis in plotly objects

Set Titles for x and y axis in plotly objects

Usage

set_axis_titles(
  plotly_object,
  curve,
  max_y_range = NA,
  min_y_range = NA,
  min_x_range = NA,
  max_x_range = NA
)

Arguments

plotly_object

a plotly object

curve

the required curve

max_y_range

the maximum value of y range (for lift and decision curve)

min_y_range

the minimum value of y range (for decision curve)

min_x_range

the minimum value of x range (for decision curve)

max_x_range

the maximum value of x range (for decision curve)