Prepare Performance Data

The prepare_performance_data function makes a Performance Data that is made of different cutoffs. Each row represents a cutoff and each column stands for a performance metric. It is possible to use this function for more than one model in order to compare different models performance for the same population. In this case the user should use a list that is made of vectors of estimated probabilities, one for each model.

Usage

prepare_performance_data(
  probs,
  real,
  by = 0.01,
  stratified_by = "probability_threshold"
)

Arguments

probs

a vector of estimated probabilities or a list of vectors of that kind (one for each model)

real

a vector of binary outcomes or a list of vectors of that kind (one for each population)

by

number: increment of the sequence.

stratified_by

Performance Metrics can be stratified by Probability Threshold or alternatively by Predicted Positives Condition Rate

Details

Sometime instead of using a cutoff for the estimated probability it is required to enforce a symmetry between the percentiles of the probabilities, in medicine it is referred as "Risk Percentile" when the outcome stands for something negative in essence such as a severe disease or death: Let's say that we want to see the model performance for the top 5% patients at risk for some well defined population, in this case the user should change the parameter stratified_by from the default "probability_threshold" to "predicted_positives" and the results will be similar Performance Data, only this time each row will represent some rounded percentile.

Examples

# You can prepare Performance Data for one model

prepare_performance_data(
  probs = example_dat$estimated_probabilities,
  real = example_dat$outcome
)
#> # A tibble: 101 × 14
#>    threshold    TP    TN    FN    FP sensitivity   FPR specificity   PPV   NPV
#>        <dbl> <int> <int> <int> <int>       <dbl> <dbl>       <dbl> <dbl> <dbl>
#>  1      0       50     0     0   100           1  1           0    0.333   NaN
#>  2      0.01    50    51     0    49           1  0.49        0.51 0.505     1
#>  3      0.02    50    55     0    45           1  0.45        0.55 0.526     1
#>  4      0.03    50    61     0    39           1  0.39        0.61 0.562     1
#>  5      0.04    50    67     0    33           1  0.33        0.67 0.602     1
#>  6      0.05    50    67     0    33           1  0.33        0.67 0.602     1
#>  7      0.06    50    67     0    33           1  0.33        0.67 0.602     1
#>  8      0.07    50    70     0    30           1  0.3         0.7  0.625     1
#>  9      0.08    50    70     0    30           1  0.3         0.7  0.625     1
#> 10      0.09    50    70     0    30           1  0.3         0.7  0.625     1
#> # … with 91 more rows, and 4 more variables: lift <dbl>,
#> #   predicted_positives <int>, NB <dbl>, ppcr <dbl>

# And you can prepare Performance Data for more than one model
prepare_performance_data(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  real = example_dat$outcome
)
#> # A tibble: 202 × 15
#>    model   threshold    TP    TN    FN    FP sensitivity   FPR specificity   PPV
#>    <chr>       <dbl> <int> <int> <int> <int>       <dbl> <dbl>       <dbl> <dbl>
#>  1 First …      0       50     0     0   100           1  1           0    0.333
#>  2 First …      0.01    50    51     0    49           1  0.49        0.51 0.505
#>  3 First …      0.02    50    55     0    45           1  0.45        0.55 0.526
#>  4 First …      0.03    50    61     0    39           1  0.39        0.61 0.562
#>  5 First …      0.04    50    67     0    33           1  0.33        0.67 0.602
#>  6 First …      0.05    50    67     0    33           1  0.33        0.67 0.602
#>  7 First …      0.06    50    67     0    33           1  0.33        0.67 0.602
#>  8 First …      0.07    50    70     0    30           1  0.3         0.7  0.625
#>  9 First …      0.08    50    70     0    30           1  0.3         0.7  0.625
#> 10 First …      0.09    50    70     0    30           1  0.3         0.7  0.625
#> # … with 192 more rows, and 5 more variables: NPV <dbl>, lift <dbl>,
#> #   predicted_positives <int>, NB <dbl>, ppcr <dbl>

# Notice that once you've put a list in the probs parameter you'll 
# receive a new
# column in the Performance Data named "Model". If it's a named list 
# (like in our
# example) the Model column will mention the names of each element 
# in the probs-list
# as the name of the model, if it's unnamed list the Models will count 
# "Model 1",
# "Model 2", etc.. according to the order of the estimated-probabilities 
# vector in the
# list.

prepare_performance_data(
  probs = list(
    "train" = example_dat %>%
      dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(estimated_probabilities)
  ),
  real = list(
    "train" = example_dat %>% dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(outcome),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(outcome)
  )
)
#> # A tibble: 202 × 15
#>    population threshold    TP    TN    FN    FP sensitivity   FPR specificity
#>    <chr>          <dbl> <int> <int> <int> <int>       <dbl> <dbl>       <dbl>
#>  1 train           0       28     0     0    68           1 1           0    
#>  2 train           0.01    28    36     0    32           1 0.471       0.529
#>  3 train           0.02    28    39     0    29           1 0.426       0.574
#>  4 train           0.03    28    43     0    25           1 0.368       0.632
#>  5 train           0.04    28    46     0    22           1 0.324       0.676
#>  6 train           0.05    28    46     0    22           1 0.324       0.676
#>  7 train           0.06    28    46     0    22           1 0.324       0.676
#>  8 train           0.07    28    48     0    20           1 0.294       0.706
#>  9 train           0.08    28    48     0    20           1 0.294       0.706
#> 10 train           0.09    28    48     0    20           1 0.294       0.706
#> # … with 192 more rows, and 6 more variables: PPV <dbl>, NPV <dbl>, lift <dbl>,
#> #   predicted_positives <int>, NB <dbl>, ppcr <dbl>