Prepare Performance Data

The prepare_performance_data function makes a Performance Data that is made of different cutoffs. Each row represents a cutoff and each column stands for a performance metric. It is possible to use this function for more than one model in order to compare different models performance for the same population. In this case the user should use a list that is made of vectors of estimated probabilities, one for each model.

Usage

prepare_performance_data(
  probs,
  reals,
  by = 0.01,
  stratified_by = "probability_threshold"
)

Arguments

probs

a list of vectors of estimated probabilities (one for each model or one for each population)

reals

a list of vectors of binary outcomes (one for each population)

by

number: increment of the sequence.

stratified_by

Performance Metrics can be stratified by Probability Threshold or alternatively by Predicted Positives Condition Rate

Details

Sometime instead of using a cutoff for the estimated probability it is required to enforce a symmetry between the percentiles of the probabilities, in medicine it is referred as "Risk Percentile" when the outcome stands for something negative in essence such as a severe disease or death: Let's say that we want to see the model performance for the top 5% patients at risk for some well defined population, in this case the user should change the parameter stratified_by from the default "probability_threshold" to "predicted_positives" and the results will be similar Performance Data, only this time each row will represent some rounded percentile.

Examples

# You can prepare Performance Data for one model

prepare_performance_data(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome)
)
#> # A tibble: 101 × 14
#>    probability…¹    TP    TN    FN    FP sensi…²   FPR speci…³   PPV   NPV  lift
#>            <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1          0       50     0     0   100       1  1       0    0.333   NaN  1   
#>  2          0.01    50    51     0    49       1  0.49    0.51 0.505     1  1.52
#>  3          0.02    50    55     0    45       1  0.45    0.55 0.526     1  1.58
#>  4          0.03    50    61     0    39       1  0.39    0.61 0.562     1  1.69
#>  5          0.04    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  6          0.05    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  7          0.06    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  8          0.07    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#>  9          0.08    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#> 10          0.09    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#> # … with 91 more rows, 3 more variables: predicted_positives <int>, NB <dbl>,
#> #   ppcr <dbl>, and abbreviated variable names ¹​probability_threshold,
#> #   ²​sensitivity, ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

prepare_performance_data(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome),
  stratified_by = "ppcr"
)
#> # A tibble: 101 × 13
#>    proba…¹  ppcr    TP    TN    FN    FP sensi…²   FPR speci…³   PPV   NPV  lift
#>      <dbl> <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1   0.996  0        0   100    50     0    0        0       1   NaN 0.667   NaN
#>  2   0.994  0.01     1   100    49     0    0.02     0       1     1 0.671     3
#>  3   0.994  0.02     1   100    49     0    0.02     0       1     1 0.671     3
#>  4   0.986  0.03     5   100    45     0    0.1      0       1     1 0.690     3
#>  5   0.982  0.04     5   100    45     0    0.1      0       1     1 0.690     3
#>  6   0.982  0.05     5   100    45     0    0.1      0       1     1 0.690     3
#>  7   0.971  0.06     9   100    41     0    0.18     0       1     1 0.709     3
#>  8   0.963  0.07    11   100    39     0    0.22     0       1     1 0.719     3
#>  9   0.952  0.08    11   100    39     0    0.22     0       1     1 0.719     3
#> 10   0.952  0.09    11   100    39     0    0.22     0       1     1 0.719     3
#> # … with 91 more rows, 1 more variable: predicted_positives <int>, and
#> #   abbreviated variable names ¹​probability_threshold, ²​sensitivity,
#> #   ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# Several Models

prepare_performance_data(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome)
)
#> # A tibble: 202 × 15
#>    model proba…¹    TP    TN    FN    FP sensi…²   FPR speci…³   PPV   NPV  lift
#>    <chr>   <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 Firs…    0       50     0     0   100       1  1       0    0.333   NaN  1   
#>  2 Firs…    0.01    50    51     0    49       1  0.49    0.51 0.505     1  1.52
#>  3 Firs…    0.02    50    55     0    45       1  0.45    0.55 0.526     1  1.58
#>  4 Firs…    0.03    50    61     0    39       1  0.39    0.61 0.562     1  1.69
#>  5 Firs…    0.04    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  6 Firs…    0.05    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  7 Firs…    0.06    50    67     0    33       1  0.33    0.67 0.602     1  1.81
#>  8 Firs…    0.07    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#>  9 Firs…    0.08    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#> 10 Firs…    0.09    50    70     0    30       1  0.3     0.7  0.625     1  1.88
#> # … with 192 more rows, 3 more variables: predicted_positives <int>, NB <dbl>,
#> #   ppcr <dbl>, and abbreviated variable names ¹​probability_threshold,
#> #   ²​sensitivity, ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

prepare_performance_data(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome),
  stratified_by = "ppcr"
)
#> # A tibble: 202 × 14
#>    model proba…¹  ppcr    TP    TN    FN    FP sensi…²   FPR speci…³   PPV   NPV
#>    <chr>   <dbl> <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl> <dbl>
#>  1 Firs…   0.996  0        0   100    50     0    0        0       1   NaN 0.667
#>  2 Firs…   0.994  0.01     1   100    49     0    0.02     0       1     1 0.671
#>  3 Firs…   0.994  0.02     1   100    49     0    0.02     0       1     1 0.671
#>  4 Firs…   0.986  0.03     5   100    45     0    0.1      0       1     1 0.690
#>  5 Firs…   0.982  0.04     5   100    45     0    0.1      0       1     1 0.690
#>  6 Firs…   0.982  0.05     5   100    45     0    0.1      0       1     1 0.690
#>  7 Firs…   0.971  0.06     9   100    41     0    0.18     0       1     1 0.709
#>  8 Firs…   0.963  0.07    11   100    39     0    0.22     0       1     1 0.719
#>  9 Firs…   0.952  0.08    11   100    39     0    0.22     0       1     1 0.719
#> 10 Firs…   0.952  0.09    11   100    39     0    0.22     0       1     1 0.719
#> # … with 192 more rows, 2 more variables: lift <dbl>,
#> #   predicted_positives <int>, and abbreviated variable names
#> #   ¹​probability_threshold, ²​sensitivity, ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# Several Populations

prepare_performance_data(
  probs = list(
    "train" = example_dat %>%
      dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat %>% dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(outcome),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(outcome)
  )
)
#> # A tibble: 202 × 15
#>    population probab…¹    TP    TN    FN    FP sensi…²   FPR speci…³   PPV   NPV
#>    <chr>         <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl> <dbl>
#>  1 train          0       28     0     0    68       1 1       0     0.292   NaN
#>  2 train          0.01    28    36     0    32       1 0.471   0.529 0.467     1
#>  3 train          0.02    28    39     0    29       1 0.426   0.574 0.491     1
#>  4 train          0.03    28    43     0    25       1 0.368   0.632 0.528     1
#>  5 train          0.04    28    46     0    22       1 0.324   0.676 0.56      1
#>  6 train          0.05    28    46     0    22       1 0.324   0.676 0.56      1
#>  7 train          0.06    28    46     0    22       1 0.324   0.676 0.56      1
#>  8 train          0.07    28    48     0    20       1 0.294   0.706 0.583     1
#>  9 train          0.08    28    48     0    20       1 0.294   0.706 0.583     1
#> 10 train          0.09    28    48     0    20       1 0.294   0.706 0.583     1
#> # … with 192 more rows, 4 more variables: lift <dbl>,
#> #   predicted_positives <int>, NB <dbl>, ppcr <dbl>, and abbreviated variable
#> #   names ¹​probability_threshold, ²​sensitivity, ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

prepare_performance_data(
  probs = list(
    "train" = example_dat %>%
      dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat %>% dplyr::filter(type_of_set == "train") %>%
      dplyr::pull(outcome),
    "test" = example_dat %>% dplyr::filter(type_of_set == "test") %>%
      dplyr::pull(outcome)
  ),
  stratified_by = "ppcr"
)
#> # A tibble: 202 × 14
#>    population probab…¹  ppcr    TP    TN    FN    FP sensi…²   FPR speci…³   PPV
#>    <chr>         <dbl> <dbl> <int> <int> <int> <int>   <dbl> <dbl>   <dbl> <dbl>
#>  1 train         0.996  0        0    68    28     0  0          0       1   NaN
#>  2 train         0.994  0.01     1    68    27     0  0.0357     0       1     1
#>  3 train         0.994  0.02     1    68    27     0  0.0357     0       1     1
#>  4 train         0.990  0.03     3    68    25     0  0.107      0       1     1
#>  5 train         0.984  0.04     4    68    24     0  0.143      0       1     1
#>  6 train         0.974  0.05     5    68    23     0  0.179      0       1     1
#>  7 train         0.958  0.06     6    68    22     0  0.214      0       1     1
#>  8 train         0.932  0.07     7    68    21     0  0.25       0       1     1
#>  9 train         0.922  0.08     7    68    21     0  0.25       0       1     1
#> 10 train         0.922  0.09     7    68    21     0  0.25       0       1     1
#> # … with 192 more rows, 3 more variables: NPV <dbl>, lift <dbl>,
#> #   predicted_positives <int>, and abbreviated variable names
#> #   ¹​probability_threshold, ²​sensitivity, ³​specificity
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names